MSI - Msi stock ui aggregation #39
- fixes

+++ app/code/Magento/Inventory/Observer/SourceItemsProcessor.php
@@ -97,8 +97,12 @@ class SourceItemsProcessor
             $sourceItemsForSave[] = $sourceItem;
             unset($sourceItemsForDelete[$sourceId]);
         }
-        $this->sourceItemSave->execute($sourceItemsForSave);
-        $this->deleteSourceItems($sourceItemsForDelete);
+        if ($sourceItemsForSave) {
+            $this->sourceItemSave->execute($sourceItemsForSave);
+        }
+        if ($sourceItemsForDelete) {
+            $this->deleteSourceItems($sourceItemsForDelete);
+        }
     }
 
     /**
